Fixed #24423 -- Reorganized trans/blocktrans tests.
